# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fulcrum Core API
  description: A comprehensive cloud infrastructure management system designed to orchestrate and monitor distributed cloud resources across multiple providers.
  version: 1.0.0
  contact:
    name: Fulcrum Project
    url: https://fulcrumproject.org
  license:
    name: Apache License 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
...
servers:
  - url: https://api.fulcrum.testudosrl.dev/api/v1
    description: Development server

# Default security for all operations
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: oauth2
      description: Role-based authentication using bearer tokens
      flows:
        implicit:
          authorizationUrl: https://api.fulcrum.testudosrl.dev/oauth/authorize
          scopes:
            admin: System administrator with unrestricted access
            participant: Participant administrator that can act as both provider and consumer
            agent: Agent role with access to specific operations
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error message
          example: "Invalid request parameters"
        details:
          type: string
          description: Additional error details
          example: "The field 'name' is required"
    properties.UUID:
      type: string
      format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"
    JSONObject:
      type: object
      additionalProperties: true
      description: A generic properties.JSON object

    # Participant schemas
...
    ParticipantResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "Test Participant"
        status:
          $ref: "#/components/schemas/ParticipantStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ParticipantStatus:
      type: string
      enum: [Enabled, Disabled]

    # Token schemas
...
    AgentStatus:
      type: string
      enum: [New, Connected, Disconnected, Error, Disabled]
...
    AgentResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "aws-agent-01"
        status:
          $ref: "#/components/schemas/AgentStatus"
        providerId:
          $ref: "#/components/schemas/properties.UUID"
        agentTypeId:
          $ref: "#/components/schemas/properties.UUID"
        tags:
          type: array
          items:
            type: string
          example: ["gpu", "high-memory", "ssd"]
        participant:
          $ref: "#/components/schemas/ParticipantResponse"
        agentType:
          $ref: "#/components/schemas/AgentTypeResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AgentTypeResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "Participant VM Agent"
        serviceTypes:
          type: array
          items:
            $ref: "#/components/schemas/ServiceTypeResponse"
          description: "Service types that this agent type can provide"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Service schemas
    ServiceStatus:
      type: string
      enum:
        [
          Creating,
          Created,
          Starting,
          Started,
          Stopping,
          Stopped,
          HotUpdating,
          ColdUpdating,
          Deleting,
          Deleted,
        ]
...
    ServiceResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "web-server-01"
        currentStatus:
          $ref: "#/components/schemas/ServiceStatus"
        targetStatus:
          $ref: "#/components/schemas/ServiceStatus"
        currentProperties:
          $ref: "#/components/schemas/JSONObject"
        targetProperties:
          $ref: "#/components/schemas/JSONObject"
        failedAction:
          $ref: "#/components/schemas/ServiceAction"
        errorMessage:
          type: string
          example: "Failed to start service: insufficient resources"
        retryCount:
          type: integer
          example: 0
        resources:
          $ref: "#/components/schemas/JSONObject"
        externalId:
          type: string
        agentId:
          $ref: "#/components/schemas/properties.UUID"
        serviceTypeId:
          $ref: "#/components/schemas/properties.UUID"
        groupId:
          $ref: "#/components/schemas/properties.UUID"
        providerId:
          $ref: "#/components/schemas/properties.UUID"
        consumerId:
          $ref: "#/components/schemas/properties.UUID"
        agent:
          $ref: "#/components/schemas/AgentResponse"
        serviceType:
          $ref: "#/components/schemas/ServiceTypeResponse"
        group:
          $ref: "#/components/schemas/ServiceGroupResponse"
        consumerParticipant:
          $ref: "#/components/schemas/ParticipantResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Service Type schemas
...
    ServiceTypeResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "VM"
        propertySchema:
          $ref: "#/components/schemas/PropertySchema"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Service Group schemas
...
    ServiceGroupResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "Web Servers"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Action types
    ServiceAction:
      type: string
      enum:
        [
          ServiceCreate,
          ServiceStart,
          ServiceStop,
          ServiceHotUpdate,
          ServiceColdUpdate,
          ServiceDelete,
        ]

    # Job schemas
    # Using the same ServiceAction enum for job actions
    JobStatus:
      type: string
      enum: [Pending, Processing, Completed, Failed]
    JobResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        action:
          $ref: "#/components/schemas/ServiceAction" # Use the common ServiceAction enum
        status:
          $ref: "#/components/schemas/JobStatus"
        agentId:
          $ref: "#/components/schemas/properties.UUID"
        serviceId:
          $ref: "#/components/schemas/properties.UUID"
        priority:
          type: integer
          example: 1
        requestData:
          $ref: "#/components/schemas/JSONObject"
        resultData:
          $ref: "#/components/schemas/JSONObject"
        errorMessage:
          type: string
          example: "Failed to create VM: insufficient resources"
        claimedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        agent:
          $ref: "#/components/schemas/AgentResponse"
        service:
          $ref: "#/components/schemas/ServiceResponse"
    CompleteJobRequest:
      type: object
      properties:
        externalId:
          type: string
        resources:
          $ref: "#/components/schemas/JSONObject"
    FailJobRequest:
      type: object
      required:
        - errorMessage
      properties:
        errorMessage:
          type: string
          example: "Failed to create VM: insufficient resources"

    # Metric schemas
    MetricEntityType:
      type: string
      enum: [Agent, Service, Resource]
    MetricTypeRequest:
      type: object
      required:
        - name
        - entityType
      properties:
        name:
          type: string
          example: "cpu_usage"
        entityType:
          $ref: "#/components/schemas/MetricEntityType"
    MetricTypeResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        entityType:
          $ref: "#/components/schemas/MetricEntityType"
        name:
          type: string
          example: "cpu_usage"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        totalItems:
          type: integer
        totalPages:
          type: integer
        pageSize:
          type: integer
        page:
          type: integer
    PropertySchema:
      type: object
      description: properties.JSON schema defining validation rules for service properties
      additionalProperties:
        $ref: "#/components/schemas/PropertyDefinition"
    PropertyDefinition:
      type: object
      properties:
        type:
          type: string
          enum: [string, integer, number, boolean, object, array]
          description: The data type of the property
        label:
          type: string
          description: Human-readable label for the property
        required:
          type: boolean
          description: Whether the property is required
          default: false
        default:
          description: Default value for the property
        validators:
          type: array
          items:
            $ref: "#/components/schemas/ValidatorDefinition"
          description: Array of validation rules
        properties:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/PropertyDefinition"
          description: Nested properties for object types
        items:
          $ref: "#/components/schemas/PropertyDefinition"
          description: Schema for array items
    ValidatorDefinition:
      type: object
      required:
        - type
        - value
      properties:
        type:
          type: string
          enum:
            [
              minLength,
              maxLength,
              pattern,
              enum,
              min,
              max,
              minItems,
              maxItems,
              uniqueItems,
            ]
          description: Type of validator
        value:
          description: Value for the validator (type depends on validator type)
...
paths:
...
  /jobs:
    get:
      summary: List jobs
      tags:
        - Jobs
      description: Retrieves a paginated list of jobs (admin route)
      x-auth-permissions:
        - role: admin
          permission: all jobs
        - role: participant
          permission: jobs related to its participant (as provider via agents or as consumer via services)
        - role: agent
          permission: jobs assigned to the agent
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: A paginated list of jobs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/JobResponse"
  /jobs/pending:
    get:
      summary: Get pending jobs
      tags:
        - Jobs
      description: Retrieves a list of pending jobs for the authenticated agent
      security:
        - BearerAuth: []
      x-auth-permissions:
        - role: admin
          permission: not authorized
        - role: participant
          permission: not authorized
        - role: agent
          permission: pending jobs assigned to the agent
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of pending jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/properties.UUID"
    get:
      summary: Get a job
      tags:
        - Jobs
      description: Retrieves a specific job by ID (admin route)
      responses:
        "200":
          description: The job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/claim:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/properties.UUID"
    post:
      summary: Claim a job
      tags:
        - Jobs
      description: Claims a job for processing by the authenticated agent
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Job claimed successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/complete:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/properties.UUID"
    post:
      summary: Complete a job
      tags:
        - Jobs
      description: Marks a job as completed with results
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteJobRequest"
      responses:
        "204":
          description: Job completed successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/fail:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/properties.UUID"
    post:
      summary: Fail a job
      tags:
        - Jobs
      description: Marks a job as failed with an error message
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FailJobRequest"
      responses:
        "204":
          description: Job marked as failed successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  #
  # Metric Type endpoints
  #
  /metric-types:
    get:
      summary: List metric types
      tags:
        - Metrics
      description: Retrieves a paginated list of metric types
      x-auth-permissions:
        - role: admin
          permission: all metric types
        - role: participant
          permission: all metric types
        - role: agent
          permission: all metric types
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: A paginated list of metric types
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/MetricTypeResponse"
    post:
      summary: Create a metric type
      tags:
        - Metrics
      description: Creates a new metric type
      x-auth-permissions:
        - role: admin
          permission: always
        - role: participant
          permission: not authorized
        - role: agent
          permission: not authorized
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetricTypeRequest"
      responses:
        "201":
          description: Metric type created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricTypeResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /metric-types/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/properties.UUID"
    get:
      summary: Get a metric type
      tags:
        - Metrics
      description: Retrieves a specific metric type by ID
      x-auth-permissions:
        - role: admin
          permission: all metric types
        - role: participant
          permission: all metric types
        - role: agent
          permission: all metric types
      responses:
        "200":
          description: The metric type details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricTypeResponse"
        "404":
          description: Metric type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update a metric type
      tags:
        - Metrics
      description: Updates an existing metric type
      x-auth-permissions:
        - role: admin
          permission: always
        - role: participant
          permission: not authorized
        - role: agent
          permission: not authorized
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "cpu_usage_updated"
      responses:
        "200":
          description: Metric type updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricTypeResponse"
        "404":
          description: Metric type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a metric type
      tags:
        - Metrics
      description: Deletes a metric type by ID
      x-auth-permissions:
        - role: admin
          permission: always
        - role: participant
          permission: not authorized
        - role: agent
          permission: not authorized
      responses:
        "204":
          description: Metric type deleted successfully
        "404":
          description: Metric type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  #
  # Metric Entry endpoints
  #
...